backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.3
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.6.0

# Redis & Caching
redis==5.2.0
hiredis==3.0.0

# Celery for background tasks
celery==5.4.0
kombu==5.4.2

# HTTP Client
httpx==0.27.2

# Web Scraping & Parsing
beautifulsoup4==4.12.3
lxml==5.3.0
html5lib==1.1

# PDF Generation
reportlab==4.2.5
weasyprint==62.3

# External APIs
anthropic==0.39.0
stripe==11.2.0
sendgrid==6.11.0
boto3==1.35.62

# Utilities
python-slugify==8.0.4
pydantic==2.10.2
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

# Linting & Formatting
ruff==0.8.2

# Monitoring
sentry-sdk[fastapi]==2.18.0
